only issue appliance deletion from launcher

+++ docker/datacommons-base/base-start.sh
@@ -93,7 +93,7 @@ echo "finished sync"
 #fusermount -u ${IRODS_MOUNT}
 
 # auto-delete self if it seems part of an appliance
-if [ ! -z "${PIVOT_URL}" ] && [ ! -z "${WORKFLOW_NAME}" ]; then
+if [ "$1" == "_toil_exec" ] && [ ! -z "${PIVOT_URL}" ] && [ ! -z "${WORKFLOW_NAME}" ]; then
     curl -X DELETE "${PIVOT_URL}/${WORKFLOW_NAME}"
 fi
